learned_codec/requirements.txt
torch>=2.0
torchvision
onnx
onnxruntime
Pillow
numpy
scikit-image
